requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
pandas==2.0.3
pyyaml==6.0.1
Pillow==10.0.0
openpyxl==3.1.2